Moves common files to common folder
