+++ src/main/java/redox/datamodel/media/common/Address.java
@@ -1,5 +1,5 @@
 
-package redox.datamodel.media.newmedia;
+package redox.datamodel.media.common;
 
 import java.util.HashMap;
 import java.util.Map;
