+++ src/main/java/redox/datamodel/media/common/Transmission.java
@@ -1,5 +1,5 @@
 
-package redox.datamodel.media.newmedia;
+package redox.datamodel.media.common;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -14,7 +14,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 @JsonPropertyOrder({
     "ID"
 })
-public class Message {
+public class Transmission {
 
     @JsonProperty("ID")
     private Integer iD;
